Cleanup and docs
